requirements.txt
mock
nose
pytest
pytest-cov
psycopg2-binary
testing.postgresql
